finish feature

+++ internal/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	Id       int64
+	Id       int
 	Login    string
 	Email    string
 	Password string
